Representant computation
A first test for A=K[[t1,t2]]
Good odds that the code does not function well over padics
Code is not correct if K is already a field extension (as GF(9)).

No more group value but a basis of generators (applying valuation to it provide the previous group value).